# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cakra Prima",
        "description": "Dokumentasi API Kas & Keuangan Cakra Prima",
        "version": "1.0.0"
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Daftar Akun Kas",
                "operationId": "13efa5706d1fae14fc1cf87c9f6b1427",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Tambah Akun Kas",
                "operationId": "e974666bc6c050296a8ee09a3ab6533f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BCA"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "rek"
                                    },
                                    "initial_balance": {
                                        "type": "number",
                                        "example": 1000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Detail Akun Kas",
                "operationId": "552b29d7944dfff7a0c49abc83f44ea5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update Akun Kas",
                "operationId": "9be8085be375b01725706b78df130bca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "initial_balance": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Hapus Akun Kas",
                "operationId": "102c60c5fd1e94be8c71c4efbeb51bf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
        "/api/cash-flow-periods/{period}/budget-needs": {
            "post": {
                "tags": [
                    "Budget Needs"
                ],
                "summary": "Tambah Kebutuhan Anggaran",
                "operationId": "91d19339fcf0d082a763888e93ef2f53",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "amount"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Tagihan Listrik"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/cash-flow-periods/{period}/budget-needs/{budgetNeed}": {
            "put": {
                "tags": [
                    "Budget Needs"
                ],
                "summary": "Update Kebutuhan Anggaran",
                "operationId": "05bde28a826016bb523bc25bab3eacc9",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "budgetNeed",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Budget Needs"
                ],
                "summary": "Hapus Kebutuhan Anggaran",
                "operationId": "34b5dd0823a43b04b8668634fd72e348",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "budgetNeed",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}